AssertTokens Trait
